refactor : highlightText 함수 분리
